Database connection and its associated files are created. Package names updated. A input form created to test Post methods on Handler Database CURD operations created
