+++ WEB_ROOT/index.html
@@ -3,9 +3,16 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Test Page</title>
+    <title>Sample Form</title>
 </head>
 <body>
-<h1>Hello, World!</h1>
+    <h1>Submit Information</h1>
+    <form action="/submit" method="post">
+        <label for="name">Name:</label>
+        <input type="text" id="name" name="name"><br><br>
+        <label for="value">Value:</label>
+        <input type="text" id="value" name="value"><br><br>
+        <button type="submit">Submit</button>
+    </form>
 </body>
 </html>
