+++ pom.xml
@@ -50,6 +50,12 @@
             <artifactId>javac</artifactId>
             <version>9+181-r4173-1</version>
         </dependency>
+        <dependency>
+            <groupId>mysql</groupId>
+            <artifactId>mysql-connector-java</artifactId>
+            <version>8.0.29</version>
+        </dependency>
+
 
     </dependencies>
