+++ src/main/java/com/ues/database/DatabaseConfig.java
@@ -1,0 +1,29 @@
+package com.ues.database;
+
+import java.io.InputStream;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.util.Properties;
+
+public class DatabaseConfig {
+
+    private static final Properties properties = new Properties();
+
+    static {
+        try (InputStream input = DatabaseConfig.class.getClassLoader().getResourceAsStream("application.properties")) {
+            properties.load(input);
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            throw new RuntimeException("Error loading database properties", ex);
+        }
+    }
+
+    public static Connection getConnection() throws SQLException {
+        return DriverManager.getConnection(
+            properties.getProperty("db.url"),
+            properties.getProperty("db.user"),
+            properties.getProperty("db.password")
+        );
+    }
+}
