+++ src/main/java/com/ues/database/ResourceManager.java
@@ -1,0 +1,53 @@
+package com.ues.database;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.util.Map;
+
+public class ResourceManager {
+
+    public static boolean createResource(Map<String, String> data) {
+        String sql = "INSERT INTO Submissions (name, value) VALUES (?, ?)";
+        try (Connection conn = DatabaseConfig.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, data.get("name"));
+            stmt.setString(2, data.get("value"));
+            int result = stmt.executeUpdate();
+            return result > 0;
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+
+    public static boolean updateResource(String id, Map<String, String> data) {
+        String sql = "UPDATE resources SET name = ?, value = ? WHERE id = ?";
+        try (Connection conn = DatabaseConfig.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, data.get("name"));
+            stmt.setString(2, data.get("value"));
+            stmt.setInt(3, Integer.parseInt(id));
+            int result = stmt.executeUpdate();
+            return result > 0;
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+
+    public static boolean deleteResource(String id) {
+        String sql = "DELETE FROM resources WHERE id = ?";
+        try (Connection conn = DatabaseConfig.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setInt(1, Integer.parseInt(id));
+            int result = stmt.executeUpdate();
+            return result > 0;
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+    
+    
+}
