+++ src/main/java/com/ues/http/HttpStatusMapping.java
@@ -1,4 +1,4 @@
-package com.example.http;
+package com.ues.http;
 
 import java.util.HashMap;
 import java.util.Map;
