+++ src/main/resources/application.properties
@@ -1,2 +1,8 @@
 spring.security.user.name=admin
 spring.security.user.password=admin
+
+# Database Configuration
+db.url=jdbc:mysql://localhost:3306/FormData
+db.user=root
+db.password=rootuser
+db.driver=com.mysql.cj.jdbc.Driver
